Moved calculation of total cart cost from cart route to the cart state in the redux store as the checkout route needs the total cost as well. The function to calculate cost was moved from the cartItems component to helperfunctions and cart reducers were updated to recalculate total cost as an item is added/removed/deleted. onClick event for button to reduce item qty updated to delete the item if current qty for item is 1.
