+++ src/Routes/Checkout/Components/TipButton/TipButton.tsx
@@ -1,0 +1,5 @@
+const TipButton = () => {
+  return <></>;
+};
+
+export default TipButton;
