+++ src/Routes/Checkout/Components/TipButton/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./TipButton";
